refactor(pptx): extract text.rs from lib.rs (verbatim)
Text body/paragraph/run parsing (parse_text_body, ShapeKind, math_run_size, math_run_color, push_math_runs, parse_paragraph, parse_bullet, parse_run) plus the list-style level-* helpers (LevelFontSizes/LevelIndent/LevelIndents/LevelBullets types + read/extract/has/merge_level_* + extract_lvl1_font_size + empty_level_bullets), now pub(crate). The master extractors that also use the level-* helpers stay in lib.rs and resolve them via use text::*. text.rs imports parse_color_node/parse_shadow (fill), resolve_theme_typeface (theme), PptxZip + shared XML helpers (crate, now pub(crate)), mime_from_ext/parse_omath_nodes/parse_lnspc/SpaceLine (ooxml_common). resolve_path + PptxZip marked pub(crate); dead parse_omath_nodes/parse_lnspc/SpaceLine imports removed from lib.rs.
